Move error structs to new mod
